feat(AuthenticationConfig): 인증이 불필요한 경로 설정 변경
- 회원가입 및 로그인 경로(/member/auth)만 인증이 불필요하도록 변경
